+++ src/DoubleConstants.cs
@@ -181,16 +181,6 @@ public static class DoubleConstants
     public const double k = BoltzmannConstant;
 
     /// <summary>
-    /// The specific heat multiplied by the specific gas constant of dry air at constant pressure, in SI base units.
-    /// </summary>
-    public const double CpTimesRSpecificDryAir = CpDryAir * RSpecificDryAir;
-
-    /// <summary>
-    /// The heat of vaporization of water, squared, in SI base units.
-    /// </summary>
-    public const double DeltaHvapWaterSquared = (double)DeltaHvapWater * DeltaHvapWater;
-
-    /// <summary>
     /// The mass of a electron, in kg.
     /// </summary>
     public const double ElectronMass = 9.10938356e-31;
@@ -205,11 +195,6 @@ public static class DoubleConstants
     public const double qe = ElementaryCharge;
 
     /// <summary>
-    /// Four times The Stefan–Boltzmann constant, in SI base units.
-    /// </summary>
-    public const double FourSigma = 4 * sigma;
-
-    /// <summary>
     /// The gravitational constant, in SI base units.
     /// </summary>
     public const double GravitationalConstant = 6.67408e-11;
@@ -217,10 +202,6 @@ public static class DoubleConstants
     /// The gravitational constant, in SI base units.
     /// </summary>
     public const double G = GravitationalConstant;
-    /// <summary>
-    /// Twice the gravitational constant, in SI base units.
-    /// </summary>
-    public const double TwoG = 2 * G;
 
     /// <summary>
     /// The heat of vaporization of water, in SI base units.
@@ -250,11 +231,6 @@ public static class DoubleConstants
     public const double MAir = MolarMassOfAir;
 
     /// <summary>
-    /// The molar mass of air divided by the universal gas constant, in SI base units.
-    /// </summary>
-    public const double MAirOverR = MAir / R;
-
-    /// <summary>
     /// The mass of a neutron, in kg.
     /// </summary>
     public const double NeutronMass = 1.674927471e-27;
@@ -272,16 +248,6 @@ public static class DoubleConstants
     /// The mass of a proton, in kg.
     /// </summary>
     public const double ProtonMass = 1.672621898e-27;
-
-    /// <summary>
-    /// The specific gas constant divided by the specific heat of dry air at constant pressure, in SI base units.
-    /// </summary>
-    public const double RSpecificOverCpDryAir = RSpecificDryAir / CpDryAir;
-
-    /// <summary>
-    /// The ratio of the specific gas constants of dry air to water, in SI base units.
-    /// </summary>
-    public const double RSpecificRatioOfDryAirToWater = RSpecificDryAir / RSpecificWater;
 
     /// <summary>
     /// The specific gas constant of dry air, in SI base units.